Move C086 numeric shape checks into facts
